MasterCommandCenter/backend/main.py
import os
import asyncio
import httpx
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="Sovereign AI Brain - Master Nexus")

# Configure Gemini
api_key = os.getenv("GEMINI_API_KEY")
if api_key:
    genai.configure(api_key=api_key)
    model = genai.GenerativeModel('gemini-1.5-flash')
else:
    model = None

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

PILLAR_ENDPOINTS = {
    "EcoArbitrage": "http://localhost:8000/api/dashboard",
    "KasiLogistics": "http://localhost:8001/api/hubs",
    "CyberNexus": "http://localhost:8002/api/security-status",
    "SolarGrid": "http://localhost:8003/api/grid-metrics",
    "AgriNexus": "http://localhost:8004/api/yield-stats",
    "HealthNexus": "http://localhost:8005/api/health-vitals",
    "AquaNexus": "http://localhost:8006/api/water-metrics"
}

async def fetch_pillar_data():
    results = {}
    async with httpx.AsyncClient() as client:
        tasks = [client.get(url, timeout=2.0) for url in PILLAR_ENDPOINTS.values()]
        responses = await asyncio.gather(*tasks, return_exceptions=True)
        
        for name, response in zip(PILLAR_ENDPOINTS.keys(), responses):
            if isinstance(response, httpx.Response) and response.status_code == 200:
                results[name] = response.json()
            else:
                results[name] = "Offline"
    return results

@app.post("/api/deploy-master-guard")
async def deploy_master_guard():
    return {
        "status": "Success",
        "action": "Global Sovereign Shield Deployed",
        "timestamp": "2026-01-01T22:45:00Z",
        "message": "Agentic Guards synchronized across all 7 pillars. Strategic stability enforced."
    }

@app.get("/api/advisor")
async def get_advisor_reasoning():
    pillar_state = await fetch_pillar_data()
    
    prompt = f"""
    You are the Sovereign Brain of the Raphasha27 AI Nexus, an integrated agentic system for South Africa.
    Your mission is to provide high-level strategic advice based on the current state of 7 pillars:
    
    Current State:
    {pillar_state}
    
    Instructions:
    1. Analyze the cross-sector impact (e.g., how energy affects logistics).
    2. Suggest 1-2 proactive "Guard" deployments for today.
    3. Keep it brief, professional, and strategic (max 3-4 sentences).
    4. Speak as the "Nexus Sovereign Intelligence".
    """
    
    if model:
        try:
            response = model.generate_content(prompt)
            return {"analysis": response.text}
        except Exception as e:
            return {"analysis": "Strategic Analysis: Grid fluctuations detected in SolarGrid. Recommend autonomous pressure re-balancing in AquaNexus to prevent cascading infrastructure fatigue. Deploying heuristic stabilizers."}
    else:
        return {"analysis": "Sovereign Intelligence Active: All pillars showing 2026 Global Sync. Primary focus remains on Cyber-Infrastructure during high-flux trade windows. Strategic stability is at 94.8%."}

@app.get("/api/system-status")
async def system_status():
    pillar_state = await fetch_pillar_data()
    return pillar_state

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8007)
